Add files via upload
This portfolio showcases my hands-on experience and analytical approach to incident response, guided by the NIST Cybersecurity Framework (CSF). It includes a detailed breakdown of simulated and real-world-inspired cybersecurity incidents, from detection to recovery. The goal of this project is to demonstrate my ability to identify threats, assess vulnerabilities, respond effectively, and propose remediation strategies in alignment with industry standards.